WIP: Fix the way of getting  Instagram Account ID The way Facebook allow to get Instagram Account ID has changed. Now for the user to be able to use with this lib their Instagram Accounts has to meet to requirements: 1. Be an Instagram Business Account 2. Be this Instagram Business Account linked to a Facebook page.
This is super annoying and a terrible DX and UX experience. But is the
price to play with a big corp I guess.

So this PR try to fix this lib by adding 2 new endpoints and removing
the old `/me` which is not used anymore.
The new endpoints are
1. `/me/accounts` to get a list of authorized FB pages by the user
   (accessToken)
2. /${facebook-page-id} to get the linked Instagram Business account of
   that page.

This involves some UI thinking. This is what I think I would do:
1. Prompt users to log in with Facebook. Set the needed API permissions
2. When they log in with Facebook and accept necesary permissions I show
   then the name of the pages they authorized. They chose one
3. I do a second request to get the linked Instagram Business Account

Now is when we can start publishing content on behalve of the user.